backend/requirements.txt
# Core Flask API
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-jwt-extended==4.5.2
flask-bcrypt==1.0.1

# Basic utilities
requests==2.31.0
python-dotenv==1.0.0

# GIS & Data Processing
# Note for Windows users: Install wheels for these packages manually if pip fails.
# See: https://www.lfd.uci.edu/~gohlke/pythonlibs/
# Use versions compatible with mapclassify 2.10.0
# mapclassify 2.10.0 requires: numpy>=1.26, pandas>=2.1, scikit-learn>=1.4, scipy>=1.12
numpy>=1.26
rasterio==1.3.8
geopandas==0.13.2
shapely==2.0.1
pyproj==3.6.0
matplotlib==3.7.2
scikit-learn>=1.4
pillow==10.0.0

# Satellite data access
sentinelsat>=1.2.1,<2.0.0
